+++ patches/src/main/kotlin/app/revanced/patches/spotify/misc/lyrics/ChangeLyricsProviderPatch.kt
@@ -6,7 +6,6 @@ import app.revanced.patcher.extensions.InstructionExtensions.replaceInstruction
 import app.revanced.patcher.patch.bytecodePatch
 import app.revanced.patcher.patch.stringOption
 import app.revanced.patcher.util.proxy.mutableTypes.MutableMethod
-import app.revanced.patches.spotify.shared.IS_SPOTIFY_LEGACY_APP_TARGET
 import app.revanced.util.getReference
 import app.revanced.util.indexOfFirstInstructionOrThrow
 import app.revanced.util.indexOfFirstInstructionReversedOrThrow
@@ -57,16 +56,10 @@ val changeLyricsProviderPatch = bytecodePatch(
     }
 
     execute {
-        if (IS_SPOTIFY_LEGACY_APP_TARGET) {
-            Logger.getLogger(this::class.java.name).severe(
-                "Change lyrics provider patch is not supported for this target version."
-            )
-            return@execute
-        }
-
         val httpClientBuilderMethod = httpClientBuilderFingerprint.originalMethod
 
         // region Create a modified copy of the HTTP client builder method with the custom lyrics provider host.
+
         val patchedHttpClientBuilderMethod = with(httpClientBuilderMethod) {
             val invokeBuildUrlIndex = indexOfFirstInstructionOrThrow {
                 getReference<MethodReference>()?.returnType == "Lokhttp3/HttpUrl;"
@@ -89,9 +82,11 @@ val changeLyricsProviderPatch = bytecodePatch(
                 httpClientBuilderFingerprint.classDef.methods.add(this)
             }
         }
+
         //endregion
 
         // region Replace the call to the HTTP client builder method used exclusively for lyrics by the modified one.
+
         getLyricsHttpClientFingerprint(httpClientBuilderMethod).method.apply {
             val getLyricsHttpClientIndex = indexOfFirstInstructionOrThrow {
                 getReference<MethodReference>() == httpClientBuilderMethod
@@ -118,6 +113,7 @@ val changeLyricsProviderPatch = bytecodePatch(
                 )
             )
         }
+
         //endregion
     }
 }